requirements.txt
absl-py==0.9.0
altgraph==0.16.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.4.1
macholib==1.11
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
modulegraph==0.17
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.2
py2app==0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.13.1
PyQt5-sip==12.7.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==0.16.0
wrapt==1.12.1
zipp==3.1.0
